// nextjs-frontend/app/openapi-client/types.gen.ts
export type BearerResponse = {
  /**
   * Access Token
   */
  access_token: string;
  /**
   * Token Type
   */
  token_type: string;
};

/**
 * Body_auth-reset:forgot_password
 */
export type BodyAuthResetForgotPassword = {
  /**
   * Email
   */
  email: string;
};

/**
 * Body_auth-reset:reset_password
 */
export type BodyAuthResetResetPassword = {
  /**
   * Token
   */
  token: string;
  /**
   * Password
   */
  password: string;
};
// ...
export type ItemCreate = {
  /**
   * Name
   */
  name: string;
  /**
   * Description
   */
  description?: string | null;
  /**
   * Quantity
   */
  quantity?: number | null;
};

/**
 * ItemRead
 */
export type ItemRead = {
  /**
   * Name
   */
  name: string;
  /**
   * Description
   */
  description?: string | null;
  /**
   * Quantity
   */
  quantity?: number | null;
  /**
   * Id
   */
  id: string;
  /**
   * User Id
   */
  user_id: string;
};

/**
 * Page[ItemRead]
 */
export type PageItemRead = {
  /**
   * Items
   */
  items: Array<ItemRead>;
  /**
   * Total
   */
  total?: number | null;
  /**
   * Page
   */
  page: number | null;
  /**
   * Size
   */
  size: number | null;
  /**
   * Pages
   */
  pages?: number | null;
};

/**
 * UserCreate
 */
export type UserCreate = {
  /**
   * Email
   */
  email: string;
  /**
   * Password
   */
  password: string;
  /**
   * Is Active
   */
  is_active?: boolean | null;
  /**
   * Is Superuser
   */
  is_superuser?: boolean | null;
  /**
   * Is Verified
   */
  is_verified?: boolean | null;
};

/**
 * UserRead
 */
export type UserRead = {
  /**
   * Id
   */
  id: string;
  /**
   * Email
   */
  email: string;
  /**
   * Is Active
   */
  is_active?: boolean;
  /**
   * Is Superuser
   */
  is_superuser?: boolean;
  /**
   * Is Verified
   */
  is_verified?: boolean;
};
// ...
export type Login = {
  /**
   * Grant Type
   */
  grant_type?: string | null;
  /**
   * Username
   */
  username: string;
  /**
   * Password
   */
  password: string;
  /**
   * Scope
   */
  scope?: string;
  /**
   * Client Id
   */
  client_id?: string | null;
  /**
   * Client Secret
   */
  client_secret?: string | null;
};

export type AuthJwtLoginData = {
  body: Login;
  path?: never;
  query?: never;
  url: "/auth/jwt/login";
};
// ...
export type AuthJwtLogoutData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/auth/jwt/logout";
};

export type AuthJwtLogoutErrors = {
  /**
   * Missing token or inactive user.
   */
  401: unknown;
};

export type AuthJwtLogoutResponses = {
  /**
   * Successful Response
   */
  200: unknown;
};

export type RegisterRegisterData = {
  body: UserCreate;
  path?: never;
  query?: never;
  url: "/auth/register";
};
// ...
export type ResetForgotPasswordData = {
  body: BodyAuthResetForgotPassword;
  path?: never;
  query?: never;
  url: "/auth/forgot-password";
};

export type ResetForgotPasswordErrors = {
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type ResetForgotPasswordError =
  ResetForgotPasswordErrors[keyof ResetForgotPasswordErrors];

export type ResetForgotPasswordResponses = {
  /**
   * Successful Response
   */
  202: unknown;
};

export type ResetResetPasswordData = {
  body: BodyAuthResetResetPassword;
  path?: never;
  query?: never;
  url: "/auth/reset-password";
};
// ...
export type UsersCurrentUserData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/users/me";
};
// ...
export type ReadItemData = {
  body?: never;
  path?: never;
  query?: {
    /**
     * Page
     * Page number
     */
    page?: number;
    /**
     * Size
     * Page size
     */
    size?: number;
  };
  url: "/items/";
};

export type ReadItemErrors = {
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type ReadItemError = ReadItemErrors[keyof ReadItemErrors];

export type ReadItemResponses = {
  /**
   * Successful Response
   */
  200: PageItemRead;
};

export type ReadItemResponse = ReadItemResponses[keyof ReadItemResponses];

export type CreateItemData = {
  body: ItemCreate;
  path?: never;
  query?: never;
  url: "/items/";
};

export type CreateItemErrors = {
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type CreateItemError = CreateItemErrors[keyof CreateItemErrors];

export type CreateItemResponses = {
  /**
   * Successful Response
   */
  200: ItemRead;
};

export type CreateItemResponse = CreateItemResponses[keyof CreateItemResponses];

export type DeleteItemData = {
  body?: never;
  path: {
    /**
     * Item Id
     */
    item_id: string;
  };
  query?: never;
  url: "/items/{item_id}";
};